Se añade la clase YOPLPAudioPlayer y se crean las funciones básicas de reproducción. Se comprueba que se pierden parte de los formatos que se pueden reproducir con la clase MediaPlayer.

+++ app/src/main/java/es/ait/yoplp/message/NewTimeMessage.java
@@ -8,14 +8,14 @@ import es.ait.yoplp.Utils;
 public class NewTimeMessage
 {
 
-    private int newTime;
+    private long newTime;
 
-    public NewTimeMessage( int newTime )
+    public NewTimeMessage( long newTime )
     {
         this.newTime = newTime;
     }
 
-    public int getNewTime()
+    public long getNewTime()
     {
         return newTime;
     }
